Move legacy class
