+++ tests/Unit/EventVisualizerTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace JonasPardon\LaravelEventVisualizer\Tests\Unit;
 
 use Illuminate\Support\Facades\Config;
-use JonasPardon\LaravelEventVisualizer\EventVisualizer;
+use JonasPardon\LaravelEventVisualizer\EventVisualizerLegacy;
 use JonasPardon\LaravelEventVisualizer\Models\Event;
 use JonasPardon\LaravelEventVisualizer\Models\Listener;
 use JonasPardon\LaravelEventVisualizer\Services\CodeParserLegacy;
@@ -31,7 +31,7 @@ final class EventVisualizerTest extends TestCase
     public function it_parses_basic_events(): void
     {
         $this->markTestSkipped('WIP');
-        $visualizer = new EventVisualizer($this->codeParser);
+        $visualizer = new EventVisualizerLegacy($this->codeParser);
 
         $event = new Event('App\\Events\\Event1');
         $listener1 = new Listener('App\\Listeners\\Listener1');
@@ -60,7 +60,7 @@ final class EventVisualizerTest extends TestCase
         $this->markTestSkipped('WIP');
         Config::set('event-visualizer.show_laravel_events', true);
 
-        $visualizer = new EventVisualizer($this->codeParser);
+        $visualizer = new EventVisualizerLegacy($this->codeParser);
 
         $appEvent = new Event('App\\Events\\Event');
         $laravelEvent = new Event('Illuminate\\Auth\\Events\\Login');
@@ -91,7 +91,7 @@ final class EventVisualizerTest extends TestCase
         $this->markTestSkipped('WIP');
         Config::set('event-visualizer.show_laravel_events', false);
 
-        $visualizer = new EventVisualizer($this->codeParser);
+        $visualizer = new EventVisualizerLegacy($this->codeParser);
 
         $appEvent = new Event('App\\Events\\Event');
         $laravelEvent = new Event('Illuminate\\Auth\\Events\\Login');
@@ -124,7 +124,7 @@ final class EventVisualizerTest extends TestCase
             'ListenerToIgnore',
         ]);
 
-        $visualizer = new EventVisualizer($this->codeParser);
+        $visualizer = new EventVisualizerLegacy($this->codeParser);
 
         $event = new Event('App\\Events\\Event');
         $listenerToInclude = new Listener('App\\Listeners\\Listener');
@@ -156,7 +156,7 @@ final class EventVisualizerTest extends TestCase
             'EventToIgnore',
         ]);
 
-        $visualizer = new EventVisualizer($this->codeParser);
+        $visualizer = new EventVisualizerLegacy($this->codeParser);
 
         $eventToInclude = new Event('App\\Events\\Event');
         $eventToIgnore = new Event('App\\Events\\EventToIgnore');
